+++ src/gui.js
@@ -5239,9 +5239,9 @@ IDE_Morph.prototype.aboutSnap = function () {
         + '\nAchal Dave: Web Audio'
         + '\nJoe Otto: Morphic Testing and Debugging'
         + '\n\n'
-        + 'Jahrd, Derec, and Jamet costumes are watercolor paintings'
-        + '\nby Meghan Taylor and represent characters from her'
-        + '\nwebcomic Prophecy of the Circle, licensed to us only'
+        + 'Jahrd, Derec, Jamet and Sarron costumes are watercolor'
+        + '\npaintings by Meghan Taylor and represent characters from'
+        + '\nher webcomic Prophecy of the Circle, licensed to us only'
         + '\nfor use in Snap! projects. Meghan also painted the Tad'
         + '\ncostumes, but that character is in the public domain.';